# Replace floating point from_chars with newer fast_float