--- a/include/boost/charconv/detail/config.hpp
+++ b/include/boost/charconv/detail/config.hpp
@@ -193,9 +193,11 @@ static_assert((BOOST_CHARCONV_ENDIAN_BIG_BYTE || BOOST_CHARCONV_ENDIAN_LITTLE_BY
 
 #ifdef __has_include
 #  if  __has_include(<bit>)
-#    include <bit>
-#    if __cpp_lib_bit_cast >= 201806L
-#       define BOOST_CHARCONV_HAS_STD_BITCAST
+#    if __cplusplus >= 202002L || BOOST_MSVC >= 202002L
+#      include <bit>
+#      if __cpp_lib_bit_cast >= 201806L
+#         define BOOST_CHARCONV_HAS_STD_BITCAST
+#      endif
 #    endif
 #  endif
 #endif