See: https://github.com/fastfloat/fast_float
As of version 5 it is licensed under BSL.