@mborland 2023-05-31
Here are the benchmarks from `benchmarks/from_chars_floating.cpp`:

On current develop:
```
GNU C++ version 13.1.0
GNU libstdc++ version 20230426

                std::strtox<float>, scientific:   755 ms (s=-2.83944e+20)
            std::from_chars<float>, scientific:   731 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, scientific:  2696 ms (s=-2.83944e+20)

                std::strtox<double>, scientific:   919 ms (s=1.47618e+299)
            std::from_chars<double>, scientific:   715 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, scientific:  3724 ms (s=1.47618e+299)

                std::strtox<float>, general:   746 ms (s=-2.83944e+20)
            std::from_chars<float>, general:   704 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, general:  2620 ms (s=-2.83944e+20)

                std::strtox<double>, general:   912 ms (s=1.47618e+299)
            std::from_chars<double>, general:   713 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, general:  3898 ms (s=1.47618e+299)

                std::strtox<float>, uint64:   788 ms (s=1.01167e+19)
            std::from_chars<float>, uint64:   916 ms (s=1.01167e+19)
boost::charconv::from_chars<float>, uint64:  3481 ms (s=1.01167e+19)

                std::strtox<double>, uint64:   767 ms (s=1.01167e+19)
            std::from_chars<double>, uint64:   912 ms (s=1.01167e+19)
boost::charconv::from_chars<double>, uint64:  3294 ms (s=1.01167e+19)
```

With this PR:

```
GNU C++ version 13.1.0
GNU libstdc++ version 20230426

                std::strtox<float>, scientific:   758 ms (s=-2.83944e+20)
            std::from_chars<float>, scientific:   730 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, scientific:  2306 ms (s=-2.83944e+20)

                std::strtox<double>, scientific:   917 ms (s=1.47618e+299)
            std::from_chars<double>, scientific:   714 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, scientific:  2441 ms (s=1.47618e+299)

                std::strtox<float>, general:   747 ms (s=-2.83944e+20)
            std::from_chars<float>, general:   705 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, general:  2208 ms (s=-2.83944e+20)

                std::strtox<double>, general:   909 ms (s=1.47618e+299)
            std::from_chars<double>, general:   710 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, general:  2448 ms (s=1.47618e+299)

                std::strtox<float>, uint64:   749 ms (s=1.01167e+19)
            std::from_chars<float>, uint64:   869 ms (s=1.01167e+19)
boost::charconv::from_chars<float>, uint64:  3149 ms (s=1.01167e+19)

                std::strtox<double>, uint64:   734 ms (s=1.01167e+19)
            std::from_chars<double>, uint64:   871 ms (s=1.01167e+19)
boost::charconv::from_chars<double>, uint64:  3150 ms (s=1.01167e+19)
```