@mborland 2023-05-31
Benchmarks from `benchmarks/from_chars_floating.cpp`. Run on an M1 Mac with macOS 13.4 and GCC 13:
```
GNU C++ version 13.1.0
GNU libstdc++ version 20230426

                std::strtox<float>, scientific:   607 ms (s=-2.83944e+20)
            std::from_chars<float>, scientific:   530 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, scientific:   535 ms (s=-2.83944e+20)

                std::strtox<double>, scientific:   778 ms (s=1.47618e+299)
            std::from_chars<double>, scientific:   531 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, scientific:   529 ms (s=1.47618e+299)

                std::strtox<float>, general:   597 ms (s=-2.83944e+20)
            std::from_chars<float>, general:   513 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, general:   519 ms (s=-2.83944e+20)

                std::strtox<double>, general:   770 ms (s=1.47618e+299)
            std::from_chars<double>, general:   528 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, general:   526 ms (s=1.47618e+299)

                std::strtox<float>, uint64:   571 ms (s=1.01167e+19)
            std::from_chars<float>, uint64:   703 ms (s=1.01167e+19)
boost::charconv::from_chars<float>, uint64:   713 ms (s=1.01167e+19)

                std::strtox<double>, uint64:   564 ms (s=1.01167e+19)
            std::from_chars<double>, uint64:   702 ms (s=1.01167e+19)
boost::charconv::from_chars<double>, uint64:   708 ms (s=1.01167e+19)
```

On develop:

```
GNU C++ version 13.1.0
GNU libstdc++ version 20230426

                std::strtox<float>, scientific:   607 ms (s=-2.83944e+20)
            std::from_chars<float>, scientific:   528 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, scientific:   763 ms (s=-2.83944e+20)

                std::strtox<double>, scientific:   780 ms (s=1.47618e+299)
            std::from_chars<double>, scientific:   527 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, scientific:  1024 ms (s=1.47618e+299)

                std::strtox<float>, general:   594 ms (s=-2.83944e+20)
            std::from_chars<float>, general:   512 ms (s=-2.83944e+20)
boost::charconv::from_chars<float>, general:   745 ms (s=-2.83944e+20)

                std::strtox<double>, general:   770 ms (s=1.47618e+299)
            std::from_chars<double>, general:   529 ms (s=1.47618e+299)
boost::charconv::from_chars<double>, general:  1024 ms (s=1.47618e+299)

                std::strtox<float>, uint64:   585 ms (s=1.01167e+19)
            std::from_chars<float>, uint64:   705 ms (s=1.01167e+19)
boost::charconv::from_chars<float>, uint64:   793 ms (s=1.01167e+19)

                std::strtox<double>, uint64:   566 ms (s=1.01167e+19)
            std::from_chars<double>, uint64:   706 ms (s=1.01167e+19)
boost::charconv::from_chars<double>, uint64:   782 ms (s=1.01167e+19)
```