@mborland 2023-05-31
@pdimov I think this is good for review. The fast_float implementation is generally untouched outside of changing some macros.